DO NOT USE IN MODERN DAY SCRIPTS! (Bringing back for legacy support)
To ensure older versions of the scripts that may appear in forks continue to work, I have brought back the simple "packages.txt" file that many older versions of the script rely on to to work. The file contains the gnome packages, it is recommended to not use this file in "from the ground up" scripts.